% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dzd_initData.R
\name{dzd_initData_read}
\alias{dzd_initData_read}
\title{读取对账单数据}
\usage{
dzd_initData_read(
  file_name = "data-raw/dzd_initData/赛普对账单期初数修改模板.xlsx",
  lang = "en"
)
}
\arguments{
\item{file_name}{文件名}

\item{lang}{语言}
}
\value{
返回值
}
\description{
读取对账单数据
}
\examples{
dzd_initData_read()
}
