% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dzd_initData.R
\name{dzd_initData_updateAll}
\alias{dzd_initData_updateAll}
\title{更新对账单年初数据}
\usage{
dzd_initData_updateAll(config_file = "data-raw/config/conn_erp.R", data)
}
\arguments{
\item{config_file}{配置文件}

\item{data}{数据}
}
\value{
返回值
}
\description{
更新对账单年初数据
}
\examples{
dzd_initData_updateAll()
}
