% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dzd_initData.R
\name{dzd_initData_updateOne2}
\alias{dzd_initData_updateOne2}
\title{更新对账单期初数据}
\usage{
dzd_initData_updateOne2(
  erp_token = "4D181CAB-4CE3-47A3-8F2B-8AB11BB6A227",
  FOrgNumber = "100",
  FCustName = "上海药明康德新药开发有限公司",
  Fdate = "2021-12-31",
  FARBal = 1,
  FUnInvBal = 1
)
}
\arguments{
\item{erp_token}{ERP口令}

\item{FOrgNumber}{组织代码}

\item{FCustName}{客户名称}

\item{Fdate}{日期}

\item{FARBal}{应收余额}

\item{FUnInvBal}{未票余额}
}
\value{
返回值
}
\description{
更新对账单期初数据
}
\examples{
dzd_initData_updateOne()
}
