% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_mngrCost.R
\name{item_mngrCost_read}
\alias{item_mngrCost_read}
\title{读取物料的管理成本}
\usage{
item_mngrCost_read(
  file_name = "data-raw/item_mngrCost/物料管理成本维护模板.xlsx"
)
}
\arguments{
\item{file_name}{文件名}
}
\value{
返回值
}
\description{
读取物料的管理成本
}
\examples{
item_mngrCost_read()
}
