% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_mngrCost.R
\name{item_mngrCost_update}
\alias{item_mngrCost_update}
\title{写入数据}
\usage{
item_mngrCost_update(
  config_file = "data-raw/config/conn_erp.R",
  file_name = "data-raw/item_mngrCost/物料管理成本维护模板.xlsx",
  table_name = "rds_t_material_input"
)
}
\arguments{
\item{config_file}{数据库配置文件}

\item{file_name}{文件名}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
写入数据
}
\examples{
item_mngrCost_update()
}
