% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outstock_permance.R
\name{outstock_performance_calc}
\alias{outstock_performance_calc}
\title{更新处理规则}
\usage{
outstock_performance_calc(
  config_file = "data-raw/config/conn_erp.R",
  FRuleName = "赛普生物默认销售提成规则",
  FStartDate = "",
  FEndDate = ""
)
}
\arguments{
\item{config_file}{配置}

\item{FRuleName}{规则名}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}
}
\value{
返回值
}
\description{
更新处理规则
}
\examples{
outstock_performance_calc()
}
