% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outstock_permance.R
\name{outstock_performance_query}
\alias{outstock_performance_query}
\title{奖金提成查询}
\usage{
outstock_performance_query(
  config_file = "data-raw/config/conn_erp.R",
  FRuleName = "赛普生物默认销售提成规则",
  FStartDate = "2022-01-01",
  FEndDate = "2022-01-31"
)
}
\arguments{
\item{config_file}{配置文件}

\item{FRuleName}{规则}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}
}
\value{
返回值
}
\description{
奖金提成查询
}
\examples{
outstock_performance_query()
}
