% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outstock_permance.R
\name{outstock_performance_rule_names}
\alias{outstock_performance_rule_names}
\title{规则名称参数}
\usage{
outstock_performance_rule_names(config_file = "data-raw/config/conn_erp.R")
}
\arguments{
\item{config_file}{配置文件}
}
\value{
返回值
}
\description{
规则名称参数
}
\examples{
outstock_performance_rule_names()
}
