% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outstock_permance.R
\name{outstock_performance_rule_new}
\alias{outstock_performance_rule_new}
\title{新增提成规则}
\usage{
outstock_performance_rule_new(
  config_file = "data-raw/config/conn_erp.R",
  FName = "销售提成规则4",
  FParam_x = 0.04,
  FParam_y = 0.005,
  FParam_z = 0.75,
  FStartDate = "2021-04-01",
  FEndDate = "2100-12-31"
)
}
\arguments{
\item{config_file}{配置文件}

\item{FName}{规则名称}

\item{FParam_x}{参数1}

\item{FParam_y}{参数2}

\item{FParam_z}{参数3}

\item{FStartDate}{日期1}

\item{FEndDate}{日期2}
}
\value{
返回值
}
\description{
新增提成规则
}
\examples{
outstock_performance_rule_new()
}
