% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outstock_permance.R
\name{outstock_performance_rule_query}
\alias{outstock_performance_rule_query}
\title{规则查询}
\usage{
outstock_performance_rule_query(
  config_file = "data-raw/config/conn_erp.R",
  FName = "",
  lang = "cn"
)
}
\arguments{
\item{config_file}{配置文件}

\item{FName}{规则名称}

\item{lang}{语言}
}
\value{
查询结果
}
\description{
规则查询
}
\examples{
outstock_performance_rule_query()
}
